Adapted tests to support new PageLockedArray instances.
