+++ Src/ILGPU.Tests/Configurations.txt
@@ -3,12 +3,10 @@ Arrays: Debug, Release, O2
 DebugTests: Debug, Release, O2
 DisassemblerTests: Debug, Release, O2
 EnumValues: Debug, Release, O2
-ExchangeBufferOperations: Debug, Release, O2
 FixedBuffers: Debug, Release, O2
 Indices: Debug, Release, O2
 KernelEntryPoints: Debug, Release, O2
 MemoryBufferOperations : Debug, Release, O2
-MemoryCacheOperations : Debug, Release, O2
 PageLockedMemory: Debug, Release, O2
 ProfilingMarkers: Debug, Release, O2
 SharedMemory: Debug, Release, O2
